Passando a api para o pai, ainda não é o ideal
